pyproject.toml
[project]
name = "azer-backend-common"
version = "0.1.0"
description = "Azer microservice common package"
requires-python = ">=3.14"
dependencies = [
    # Web framework
    "fastapi >=0.123.5",
    "uvicorn >=0.38.0",
    # Security & Cryptography
    "argon2-cffi >=25.1.0",
    "pyjwt >=2.9.0",
    "cryptography >=46.0.3",
    # Database & ORM
    "tortoise-orm >=0.25.0",
    "asyncpg >=0.31.0",
    "aerich >=0.9.2",
    # Caching & Rate Limiting
    "redis >=7.1.0",
    "fastapi-limiter >=0.1.6",
    # Configuration
    "pydantic-settings >=2.12.0",
    "pydantic >=2.11.4",
    # Scheduling
    "apscheduler >=3.10.4",
    # Utilities
    "user-agents >=2.2.0",
    "pyyaml >=6.0.2",
    "uuid7>=0.1.0",
]

[dependency-groups]
dev = [
    # Testing
    "pytest >=9.0.1",
    "pytest-asyncio >=1.3.0",
    "schemathesis >=3.37.0",

    # Code quality
    "ruff >=0.14.7",
    "pylint >=4.0.4",
    "mypy >=1.10.0",
    "black >=24.10.0",

    # Security scanning
    "bandit >=1.7.10",
    "safety >=3.7.0",

    # Development tools
    "toml >=0.10.2",
    "tomlkit >=0.13.2",
    "tabulate >=0.9.0",
    "pyfiglet >=1.0.2",
    "pre-commit >=3.7.0",

    # Debugging
    "ipdb >=0.13.13",
]


[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "azer_common"
module-root = "."

[tool.uv]
index = [
    {url = "https://mirrors.aliyun.com/pypi/simple/", priority = 1},
    {url = "https://pypi.org/simple/", priority = 2},
]

[tool.uv.pip]
resolution = "highest"
strict = true

# Development configuration
[tool.black]
target-version = ["py313"]  # 降级到black支持的最高版本
line-length = 120  # 与ruff的line-length一致，避免格式冲突
include = '\.pyi?$'  # 格式化py/pyi文件
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | migrations  # 排除数据库迁移文件
  | tests/fixtures  # 可选：排除测试夹具（若无需严格格式化）
)/
'''

[tool.ruff]
target-version = "py314"
line-length = 120
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "S",  # flake8-bandit（安全检查）
]
ignore = [
    "E501",  # 行长度由black处理
    "B008",  # 允许参数默认值中调用函数
    "I001",  # 禁用isort与black的导入格式冲突检查
]

# ruff 导入排序配置
[tool.ruff.isort]
profile = "black"
known-first-party = ["app"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101", "B108"]


[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
exclude = [
    "venv/*",
    "build/*",
    "dist/*",
    "migrations/*",
]
[[tool.mypy.overrides]]
module = [
    "aiohttp.*",
    "alibabacloud_*.*",
    "jinja2.*",
    "fastapi.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "conftest.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --cov=app --cov-report=term-missing"
asyncio_mode = "auto"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::UserWarning:aiohttp.*",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
